refactored preview to camerapreview
