+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [26.0.0]
+- [BreakingChange][Camera] Renamed `Preview` to `CameraPreview`
+
 ## [25.0.0]
 - [BarcodeScanner] Made sure the bar code scanning result goes through an observation process to determine the barcode with the most number of scans to the consumer.
 - [BreakingChange][BarcodeScanner] BarcodeScanner.Start now requires a `DidFindBarcodeCallback(BarcodeScanResult)` instead of `Action<Barcode>`.
