+++ src/library/DIPS.Mobile.UI/API/Builder/AppHostBuilderExtensions.cs
@@ -1,4 +1,4 @@
-using DIPS.Mobile.UI.API.Camera;
+using DIPS.Mobile.UI.API.Camera.Preview;
 using DIPS.Mobile.UI.API.Library;
 using DIPS.Mobile.UI.Components.BottomSheets;
 using DIPS.Mobile.UI.Components.Chips;
@@ -12,6 +12,7 @@ using Microsoft.Maui.LifecycleEvents;
 using SkiaSharp.Views.Maui.Controls.Hosting;
 using ButtonHandler = DIPS.Mobile.UI.Components.Buttons.ButtonHandler;
 using Button = DIPS.Mobile.UI.Components.Buttons.Button;
+using CameraPreviewHandler = DIPS.Mobile.UI.API.Camera.Preview.CameraPreviewHandler;
 using CollectionView = DIPS.Mobile.UI.Components.Lists.CollectionView;
 using CollectionViewHandler = DIPS.Mobile.UI.Components.Lists.CollectionViewHandler;
 using ContextMenuPlatformEffect = DIPS.Mobile.UI.Components.ContextMenus.ContextMenuPlatformEffect;
@@ -26,7 +27,6 @@ using Label = DIPS.Mobile.UI.Components.Labels.Label;
 using LabelHandler = DIPS.Mobile.UI.Components.Labels.LabelHandler;
 using Layout = DIPS.Mobile.UI.Effects.Layout.Layout;
 using LayoutPlatformEffect = DIPS.Mobile.UI.Effects.Layout.LayoutPlatformEffect;
-using PreviewHandler = DIPS.Mobile.UI.API.Camera.PreviewHandler;
 using ScrollView = DIPS.Mobile.UI.Components.Lists.ScrollView;
 using ScrollViewHandler = DIPS.Mobile.UI.Components.Lists.ScrollViewHandler;
 using SearchBar = DIPS.Mobile.UI.Components.Searching.SearchBar;
@@ -67,7 +67,7 @@ public static partial class AppHostBuilderExtensions
             handlers.AddHandler<Entry, EntryHandler>();
             handlers.AddHandler<Editor, EditorHandler>();
             handlers.AddHandler<BottomSheet, BottomSheetHandler>();
-            handlers.AddHandler<Preview, PreviewHandler>();
+            handlers.AddHandler<CameraPreview, CameraPreviewHandler>();
             
             AddPlatformHandlers(handlers);
         });
