+++ src/library/DIPS.Mobile.UI/API/Camera/Preview/CameraPreviewHandler.cs
@@ -1,0 +1,6 @@
+namespace DIPS.Mobile.UI.API.Camera.Preview;
+
+public partial class CameraPreviewHandler
+{
+    
+}
