+++ src/library/DIPS.Mobile.UI/API/Camera/Preview/dotnet/CameraPreviewHandler.cs
@@ -1,0 +1,8 @@
+using Microsoft.Maui.Handlers;
+
+namespace DIPS.Mobile.UI.API.Camera.Preview;
+
+public partial class CameraPreviewHandler : ContentViewHandler
+{
+    
+}
