--- src/library/DIPS.Mobile.UI/API/Camera/dotnet/PreviewHandler.cs
@@ -1,8 +1,0 @@
-using Microsoft.Maui.Handlers;
-
-namespace DIPS.Mobile.UI.API.Camera;
-
-public partial class PreviewHandler : ContentViewHandler
-{
-    
-}
